ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-chroma
langchain-openai
langchain-huggingface
sentence-transformers
chromadb
pypdf
python-docx
python-dotenv
tiktoken
pydantic
unstructured
unstructured[docx]
python-magic-bin
